web/maintainer: Move /promote into /maintainer namespace
