+++ clients/apps/web/src/pages/promote/embeds/index.tsx
@@ -1,6 +1,5 @@
 import Gatekeeper from '@/components/Dashboard/Gatekeeper/Gatekeeper'
 import LoadingScreen from '@/components/Dashboard/LoadingScreen'
-import Layout from '@/components/Layout/EmptyLayout'
 import type { NextPageWithLayout } from '@/utils/next'
 import { useRouter } from 'next/router'
 import { ReactElement, useEffect } from 'react'
@@ -8,7 +7,7 @@ import { ReactElement, useEffect } from 'react'
 const Page: NextPageWithLayout = () => {
   const router = useRouter()
   useEffect(() => {
-    router.push(`/issues`)
+    router.push(`/maintainer`)
   })
 
   return (
@@ -21,11 +20,7 @@ const Page: NextPageWithLayout = () => {
 }
 
 Page.getLayout = (page: ReactElement) => {
-  return (
-    <Gatekeeper>
-      <Layout>{page}</Layout>
-    </Gatekeeper>
-  )
+  return <Gatekeeper>{page}</Gatekeeper>
 }
 
 export default Page
