+++ clients/apps/web/src/pages/promote/issues/index.tsx
@@ -8,7 +8,7 @@ import { ReactElement, useEffect } from 'react'
 const Page: NextPageWithLayout = () => {
   const router = useRouter()
   useEffect(() => {
-    router.push(`/issues`)
+    router.push(`/maintainer`)
   })
 
   return (
